feat(nyaasi): add static classes instead of object-based ones
