+++ nyaapy/anime_site.py
@@ -1,27 +1,26 @@
 import requests
-from nyaapy import utils
 from nyaapy import torrent
+from nyaapy.parser import parse_nyaa, parse_single, parse_nyaa_rss
 
+class AnimeTorrentSite:
+    SITE = torrent.TorrentSite.NYAASI
+    URL = "https://nyaa.si"
 
-class Nyaa:
-
-    def __init__(self):
-        self.SITE = utils.TorrentSite.NYAASI
-        self.URL = "https://nyaa.si"
-
-    def last_uploads(self, number_of_results):
+    @classmethod
+    def last_uploads(self, number_of_results: int):
         r = requests.get(self.URL)
 
         # If anything up with nyaa servers let the user know.
         r.raise_for_status()
 
-        json_data = utils.parse_nyaa(
-            request_text=r.text, limit=number_of_results + 1, site=self.SITE
+        json_data = parse_nyaa(
+            request_text=r.text, limit=number_of_results, site=self.SITE
         )
 
         return torrent.json_to_class(json_data)
 
-    def search(self, keyword, **kwargs):
+    @classmethod
+    def search(self, keyword: str, **kwargs):
         base_url = self.URL
 
         user = kwargs.get("user", None)
@@ -67,28 +66,30 @@ class Nyaa:
         http_response.raise_for_status()
 
         if user:
-            json_data = utils.parse_nyaa(
-                request_text=http_response.text, limit=None, site=self.SITE
+            json_data = parse_nyaa(
+                request_text=http_response.content, limit=None, site=self.SITE
             )
         else:
-            json_data = utils.parse_nyaa_rss(
-                request_text=http_response.text, limit=None, site=self.SITE
+            json_data = parse_nyaa_rss(
+                request_text=http_response.content, limit=None, site=self.SITE
             )
 
         # Convert JSON data to a class object
         return torrent.json_to_class(json_data)
 
-    def get(self, view_id):
+    @classmethod
+    def get(self, view_id: int):
         r = requests.get(f"{self.URL}/view/{view_id}")
         r.raise_for_status()
 
-        json_data = utils.parse_single(request_text=r.text, site=self.SITE)
+        json_data = parse_single(request_text=r.content, site=self.SITE)
 
         return torrent.json_to_class(json_data)
 
-    def get_user(self, username):
+    @classmethod
+    def get_from_user(self, username):
         r = requests.get(f"{self.URL}/user/{username}")
         r.raise_for_status()
 
-        json_data = utils.parse_nyaa(request_text=r.text, limit=None, site=self.SITE)
+        json_data = parse_nyaa(request_text=r.content, limit=None, site=self.SITE)
         return torrent.json_to_class(json_data)
