+++ nyaapy/nyaasi/sukebei.py
@@ -1,0 +1,6 @@
+from nyaapy.anime_site import AnimeTorrentSite
+from nyaapy.torrent import TorrentSite
+
+class SukebeiNyaa(AnimeTorrentSite):
+    SITE = TorrentSite.SUKEBEINYAASI
+    URL = "https://sukebei.nyaa.si"
