--- tests/test_pantsu.py
@@ -1,7 +1,0 @@
-"""
-* Pantsu need some serious work
-Regular data single_torrent parser not working from other Nyaa alternatives
-Needs some work
-"""
-
-print("TODO")
